requirements.txt
asgiref==3.3.1
attrs==20.3.0
certifi==2020.12.5
chardet==4.0.0
Django==3.1.4
django-filter==2.4.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
gunicorn==20.0.4
idna==2.10
iniconfig==1.1.1
packaging==20.8
pluggy==0.13.1
psycopg2-binary==2.8.5
py==1.10.0
PyJWT==2.0.0
pyparsing==2.4.7
pytest==6.2.1
pytest-django==4.1.0
pytz==2020.4
requests==2.25.1
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.2
